(SERVER-2172) Move setup arg parsing to own file
